make sbmlPower(x::Int, y::Int) dispatch

+++ src/interpret.jl
@@ -21,7 +21,7 @@ end
 sbmlLog(x) = sbmlLog(10, x)
 sbmlLog(base, x) = log(base, x)
 
-sbmlPower(x, y::Int) = x^float(y)
+sbmlPower(x::Int, y::Int) = x^float(y)
 sbmlPower(x, y) = x^y
 sbmlRoot(x) = sqrt(x)
 sbmlRoot(power, x) = x^(1 / power)
